// src/app/shared/models/form.ts
export interface FormType {
    institution: string;
    email: string;
    phoneNumber: string;
    location: string;
    activitiesRoom: boolean;
    indoorMeetings: boolean;
    laboratories: boolean;
    proyectors: boolean;
    screens: boolean;
    speakers: boolean;
    microphones: boolean;
    meetingDate: Date;
    message: string;
}